requirements.txt
unstructured==0.15.9
langchain==0.2.16
langchain-unstructured==0.1.2
python-dotenv==1.0.1
python-magic==0.4.27
